@@ -1,11 +1,23 @@
-import React from "react";
-import "./App.scss";
-import { Main } from "./containers/Main";
+import About from "./components/About";
+import Contact from "./components/Contact";
+import Experience from "./components/Experience";
+import Home from "./components/Home";
+import NavBar from "./components/NavBar";
+import Portfolio from "./components/Portfolio";
+import SocialLinks from "./components/SocialLinks";
+import Footer from "./components/Footer"
 
 function App() {
   return (
     <div>
-      <Main />
+      <NavBar />
+      <Home />
+      <About />
+      <Portfolio />
+      <Experience />
+      <Contact />
+      <Footer />
+      <SocialLinks />
     </div>
   );
 }
